refactor(config): split journal snapshot store out of io.audit
